+++ frontend/src/Dashboard/views/index.js
@@ -1,22 +1,3 @@
-import {
-  Breadcrumbs,
-  Cards,
-  Carousels,
-  Collapses,
-  Dropdowns,
-  Forms,
-  Jumbotrons,
-  ListGroups,
-  Navbars,
-  Navs,
-  Paginations,
-  Popovers,
-  ProgressBar,
-  Switches,
-  Tables,
-  Tabs,
-  Tooltips,
-} from './Base';
 
 import { ButtonDropdowns, ButtonGroups, Buttons, BrandButtons } from './Buttons';
 import Charts from './Charts';