+++ frontend/src/components/cards/CardFiltroDetalle.js
@@ -7,39 +7,39 @@ function CardFiltroDetalle(props) {
   };
   return (
 
-              <div className="col-md-4">
-                <div className="card mb-4 shadow-sm">
-                  <img src="img/home3.jpg" className="img"></img>
+    <div className="col-md-4">
+      <div className="card mb-4 shadow-sm">
+        <img src="img/home3.jpg" className="img"></img>
 
-                  <div className="card-body">
-                    <h5 className="card-title">{props.propiedad.titulo}</h5>
-                    <p className="card-text">
-                          {props.propiedad.descripcion}
-                    </p>
-                    <strong>Precio: ${props.propiedad.precio}.000</strong>
-                    <div className="d-flex justify-content-between align-items-center">
-                      <div className="btn-group">
-                      <Link to={`/propiedad/${props.propiedad.id_propiedades}`}>
+        <div className="card-body">
+          <h5 className="card-title">{props.propiedad.titulo}</h5>
+          <p className="card-text">
+            {props.propiedad.descripcion}
+          </p>
+          <strong>Precio: ${props.propiedad.precio}</strong>
+          <div className="d-flex justify-content-between align-items-center">
+            <div className="btn-group">
+              <Link to={`/propiedad/${props.propiedad.id_propiedades}`}>
 
-                        <button
-                          type="button"
-                          className="btn btn-sm btn-outline-secondary"
-                        >
-                              Ver
+                <button
+                  type="button"
+                  className="btn btn-sm btn-outline-secondary"
+                >
+                  Ver
                         </button>
-                        </Link>
+              </Link>
 
-                      </div>
-                      <small className="text-muted">
-                        <font style={NuncaExpandirse}>
-                          <font style={NuncaExpandirse}>9 minutos</font>
-                        </font>
-                      </small>
-                    </div>
-                  </div>
-                </div>
-              </div>
-   
+            </div>
+            <small className="text-muted">
+              <font style={NuncaExpandirse}>
+                <font style={NuncaExpandirse}>9 minutos</font>
+              </font>
+            </small>
+          </div>
+        </div>
+      </div>
+    </div>
+
   );
 }